+++ test/unit/attributes-mapper-spec.js
@@ -1,6 +1,8 @@
 /* global describe, it */
+
 const AttributesMapper = require("../../server/lib/attributes-mapper");
 const assert = require("assert");
+const moment = require("moment");
 
 describe("AttributesMapper", () => {
   it("should clear nested account attributes from mappings", () => {
@@ -112,6 +114,39 @@ describe("AttributesMapper", () => {
 
     const mappedAttribs = {
       "salesforce_lead-title": "Customer Success",
+      first_name: "Thomas",
+      last_name: "Bass",
+      created_at: "1360013296",
+      email: "tb@hull.io",
+      hull_segments: ["Foo", "Bar"]
+    };
+
+    const mapper = new AttributesMapper(attribMappings);
+    const result = mapper.mapAttributesForService(user, "1360013296", "tb@hull.io");
+
+    assert.deepEqual(result, mappedAttribs);
+  });
+
+  it.only("should allow to map account name", () => {
+    const attribMappings = [
+      "email",
+      "first_name",
+      "last_name",
+      "account.name"
+    ];
+
+    const user = {
+      account: {
+        name: "Hull Inc",
+      },
+      first_name: "Thomas",
+      last_name: "Bass",
+      email: "tb@hull.io",
+      hull_segments: ["Foo", "Bar"]
+    };
+
+    const mappedAttribs = {
+      account_name: "Hull Inc",
       first_name: "Thomas",
       last_name: "Bass",
       created_at: "1360013296",